flowey: factor a shared virtio-villain test-run node
The local (xflowey virtio-villain-run) and CI consume-archive villain
jobs had drifted: the local job reinvented a stripped-down runner that
threw away JUnit, never published results, and dumped per-test logs into
the internal flowey work dir instead of a stable location, while the CI
job published properly. The two jobs were also ~80% copy-paste of each
other, so every feature had to be added twice.

Extract the common staging, artifact resolution, nextest run and result
publishing into a single test_virtio_villain node (the villain analogue
of test_nextest_vmm_tests_archive). Both jobs are now thin wrappers that
differ only in the NextestRunKind they pass -- BuildAndRun from source
locally, RunFromArchive in CI. Give the local pipeline --dir and
--ci-profile flags mirroring vmm-tests-run, so a local run now publishes
JUnit and per-test logs into a self-contained results tree under
target/virtio_villain by default. Verified end-to-end: a filtered local
run passes and lands junit + logs in the --dir.

Copilot-Session: 7b5df162-05ee-40d7-85c3-ce37b6783805
